requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.66.1
opencv-python-headless==4.8.0.76
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gdown==4.7.1
